+++ pkg/alertmanager/config/config.go
@@ -21,7 +21,6 @@ import (
 	"net/url"
 	"path/filepath"
 	"regexp"
-	"sort"
 	"strings"
 	"time"
 
@@ -30,9 +29,6 @@ import (
 	"gopkg.in/yaml.v2"
 
 	commoncfg "github.com/giantswarm/prometheus-meta-operator/pkg/prometheus/common/config"
-
-	"github.com/prometheus/alertmanager/pkg/labels"
-	"github.com/prometheus/alertmanager/timeinterval"
 )
 
 const secretToken = "<secret>"
@@ -53,7 +49,7 @@ type Secret string
 // MarshalYAML implements the yaml.Marshaler interface for Secret.
 func (s Secret) MarshalYAML() (interface{}, error) {
 	if s != "" {
-		return string(s), nil
+		return secretToken, nil
 	}
 	return nil, nil
 }
@@ -66,7 +62,7 @@ func (s *Secret) UnmarshalYAML(unmarshal func(interface{}) error) error {
 
 // MarshalJSON implements the json.Marshaler interface for Secret.
 func (s Secret) MarshalJSON() ([]byte, error) {
-	return json.Marshal(string(s))
+	return json.Marshal(secretToken)
 }
 
 // URL is a custom type that represents an HTTP or HTTPS URL and allows validation at configuration load time.
@@ -219,59 +215,15 @@ func resolveFilepaths(baseDir string, cfg *Config) {
 	for i, tf := range cfg.Templates {
 		cfg.Templates[i] = join(tf)
 	}
-
-	cfg.Global.HTTPConfig.SetDirectory(baseDir)
-	for _, receiver := range cfg.Receivers {
-		for _, cfg := range receiver.OpsGenieConfigs {
-			cfg.HTTPConfig.SetDirectory(baseDir)
-		}
-		for _, cfg := range receiver.PagerdutyConfigs {
-			cfg.HTTPConfig.SetDirectory(baseDir)
-		}
-		for _, cfg := range receiver.PushoverConfigs {
-			cfg.HTTPConfig.SetDirectory(baseDir)
-		}
-		for _, cfg := range receiver.SlackConfigs {
-			cfg.HTTPConfig.SetDirectory(baseDir)
-		}
-		for _, cfg := range receiver.VictorOpsConfigs {
-			cfg.HTTPConfig.SetDirectory(baseDir)
-		}
-		for _, cfg := range receiver.WebhookConfigs {
-			cfg.HTTPConfig.SetDirectory(baseDir)
-		}
-		for _, cfg := range receiver.WechatConfigs {
-			cfg.HTTPConfig.SetDirectory(baseDir)
-		}
-	}
-}
-
-// MuteTimeInterval represents a named set of time intervals for which a route should be muted.
-type MuteTimeInterval struct {
-	Name          string                      `yaml:"name"`
-	TimeIntervals []timeinterval.TimeInterval `yaml:"time_intervals"`
-}
-
-// UnmarshalYAML implements the yaml.Unmarshaler interface for MuteTimeInterval.
-func (mt *MuteTimeInterval) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	type plain MuteTimeInterval
-	if err := unmarshal((*plain)(mt)); err != nil {
-		return err
-	}
-	if mt.Name == "" {
-		return fmt.Errorf("missing name in mute time interval")
-	}
-	return nil
 }
 
 // Config is the top-level configuration for Alertmanager's config files.
 type Config struct {
-	Global            *GlobalConfig      `yaml:"global,omitempty" json:"global,omitempty"`
-	Route             *Route             `yaml:"route,omitempty" json:"route,omitempty"`
-	InhibitRules      []*InhibitRule     `yaml:"inhibit_rules,omitempty" json:"inhibit_rules,omitempty"`
-	Receivers         []*Receiver        `yaml:"receivers,omitempty" json:"receivers,omitempty"`
-	Templates         []string           `yaml:"templates" json:"templates"`
-	MuteTimeIntervals []MuteTimeInterval `yaml:"mute_time_intervals,omitempty" json:"mute_time_intervals,omitempty"`
+	Global       *GlobalConfig  `yaml:"global,omitempty" json:"global,omitempty"`
+	Route        *Route         `yaml:"route,omitempty" json:"route,omitempty"`
+	InhibitRules []*InhibitRule `yaml:"inhibit_rules,omitempty" json:"inhibit_rules,omitempty"`
+	Receivers    []*Receiver    `yaml:"receivers,omitempty" json:"receivers,omitempty"`
+	Templates    []string       `yaml:"templates" json:"templates"`
 
 	// original is the input from which the config was parsed.
 	original string
@@ -457,23 +409,9 @@ func (c *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	if len(c.Route.Match) > 0 || len(c.Route.MatchRE) > 0 {
 		return fmt.Errorf("root route must not have any matchers")
 	}
-	if len(c.Route.MuteTimeIntervals) > 0 {
-		return fmt.Errorf("root route must not have any mute time intervals")
-	}
 
 	// Validate that all receivers used in the routing tree are defined.
-	if err := checkReceiver(c.Route, names); err != nil {
-		return err
-	}
-
-	tiNames := make(map[string]struct{})
-	for _, mt := range c.MuteTimeIntervals {
-		if _, ok := tiNames[mt.Name]; ok {
-			return fmt.Errorf("mute time interval %q is not unique", mt.Name)
-		}
-		tiNames[mt.Name] = struct{}{}
-	}
-	return checkTimeInterval(c.Route, tiNames)
+	return checkReceiver(c.Route, names)
 }
 
 // checkReceiver returns an error if a node in the routing tree
@@ -493,29 +431,11 @@ func checkReceiver(r *Route, receivers map[string]struct{}) error {
 	return nil
 }
 
-func checkTimeInterval(r *Route, timeIntervals map[string]struct{}) error {
-	for _, sr := range r.Routes {
-		if err := checkTimeInterval(sr, timeIntervals); err != nil {
-			return err
-		}
-	}
-	if len(r.MuteTimeIntervals) == 0 {
-		return nil
-	}
-	for _, mt := range r.MuteTimeIntervals {
-		if _, ok := timeIntervals[mt]; !ok {
-			return fmt.Errorf("undefined time interval %q used in route", mt)
-		}
-	}
-	return nil
-}
-
 // DefaultGlobalConfig returns GlobalConfig with default values.
 func DefaultGlobalConfig() GlobalConfig {
-	var defaultHTTPConfig = commoncfg.DefaultHTTPClientConfig
 	return GlobalConfig{
 		ResolveTimeout: model.Duration(5 * time.Minute),
-		HTTPConfig:     &defaultHTTPConfig,
+		HTTPConfig:     &commoncfg.HTTPClientConfig{},
 
 		SMTPHello:       "localhost",
 		SMTPRequireTLS:  true,
@@ -631,7 +551,7 @@ type GlobalConfig struct {
 	SMTPAuthPassword Secret     `yaml:"smtp_auth_password,omitempty" json:"smtp_auth_password,omitempty"`
 	SMTPAuthSecret   Secret     `yaml:"smtp_auth_secret,omitempty" json:"smtp_auth_secret,omitempty"`
 	SMTPAuthIdentity string     `yaml:"smtp_auth_identity,omitempty" json:"smtp_auth_identity,omitempty"`
-	SMTPRequireTLS   bool       `yaml:"smtp_require_tls" json:"smtp_require_tls,omitempty"`
+	SMTPRequireTLS   bool       `yaml:"smtp_require_tls,omitempty" json:"smtp_require_tls,omitempty"`
 	SlackAPIURL      *SecretURL `yaml:"slack_api_url,omitempty" json:"slack_api_url,omitempty"`
 	PagerdutyURL     *URL       `yaml:"pagerduty_url,omitempty" json:"pagerduty_url,omitempty"`
 	OpsGenieAPIURL   *URL       `yaml:"opsgenie_api_url,omitempty" json:"opsgenie_api_url,omitempty"`
@@ -657,14 +577,11 @@ type Route struct {
 	GroupByStr []string          `yaml:"group_by,omitempty" json:"group_by,omitempty"`
 	GroupBy    []model.LabelName `yaml:"-" json:"-"`
 	GroupByAll bool              `yaml:"-" json:"-"`
-	// Deprecated. Remove before v1.0 release.
-	Match map[string]string `yaml:"match,omitempty" json:"match,omitempty"`
-	// Deprecated. Remove before v1.0 release.
-	MatchRE           MatchRegexps `yaml:"match_re,omitempty" json:"match_re,omitempty"`
-	Matchers          Matchers     `yaml:"matchers,omitempty" json:"matchers,omitempty"`
-	MuteTimeIntervals []string     `yaml:"mute_time_intervals,omitempty" json:"mute_time_intervals,omitempty"`
-	Continue          bool         `yaml:"continue" json:"continue,omitempty"`
-	Routes            []*Route     `yaml:"routes,omitempty" json:"routes,omitempty"`
+
+	Match    map[string]string `yaml:"match,omitempty" json:"match,omitempty"`
+	MatchRE  MatchRegexps      `yaml:"match_re,omitempty" json:"match_re,omitempty"`
+	Continue bool              `yaml:"continue,omitempty" json:"continue,omitempty"`
+	Routes   []*Route          `yaml:"routes,omitempty" json:"routes,omitempty"`
 
 	GroupWait      *model.Duration `yaml:"group_wait,omitempty" json:"group_wait,omitempty"`
 	GroupInterval  *model.Duration `yaml:"group_interval,omitempty" json:"group_interval,omitempty"`
@@ -724,21 +641,17 @@ func (r *Route) UnmarshalYAML(unmarshal func(interface{}) error) error {
 // Both alerts have to have a set of labels being equal.
 type InhibitRule struct {
 	// SourceMatch defines a set of labels that have to equal the given
-	// value for source alerts. Deprecated. Remove before v1.0 release.
+	// value for source alerts.
 	SourceMatch map[string]string `yaml:"source_match,omitempty" json:"source_match,omitempty"`
 	// SourceMatchRE defines pairs like SourceMatch but does regular expression
-	// matching. Deprecated. Remove before v1.0 release.
+	// matching.
 	SourceMatchRE MatchRegexps `yaml:"source_match_re,omitempty" json:"source_match_re,omitempty"`
-	// SourceMatchers defines a set of label matchers that have to be fulfilled for source alerts.
-	SourceMatchers Matchers `yaml:"source_matchers,omitempty" json:"source_matchers,omitempty"`
 	// TargetMatch defines a set of labels that have to equal the given
-	// value for target alerts. Deprecated. Remove before v1.0 release.
+	// value for target alerts.
 	TargetMatch map[string]string `yaml:"target_match,omitempty" json:"target_match,omitempty"`
 	// TargetMatchRE defines pairs like TargetMatch but does regular expression
-	// matching. Deprecated. Remove before v1.0 release.
+	// matching.
 	TargetMatchRE MatchRegexps `yaml:"target_match_re,omitempty" json:"target_match_re,omitempty"`
-	// TargetMatchers defines a set of label matchers that have to be fulfilled for target alerts.
-	TargetMatchers Matchers `yaml:"target_matchers,omitempty" json:"target_matchers,omitempty"`
 	// A set of labels that must be equal between the source and target alert
 	// for them to be a match.
 	Equal model.LabelNames `yaml:"equal,omitempty" json:"equal,omitempty"`
@@ -842,7 +755,7 @@ func (re Regexp) MarshalYAML() (interface{}, error) {
 	return nil, nil
 }
 
-// UnmarshalJSON implements the json.Unmarshaler interface for Regexp
+// UnmarshalJSON implements the json.Marshaler interface for Regexp
 func (re *Regexp) UnmarshalJSON(data []byte) error {
 	var s string
 	if err := json.Unmarshal(data, &s); err != nil {
@@ -863,63 +776,4 @@ func (re Regexp) MarshalJSON() ([]byte, error) {
 		return json.Marshal(re.original)
 	}
 	return nil, nil
-}
-
-// Matchers is label.Matchers with an added UnmarshalYAML method to implement the yaml.Unmarshaler interface
-// and MarshalYAML to implement the yaml.Marshaler interface.
-type Matchers labels.Matchers
-
-// UnmarshalYAML implements the yaml.Unmarshaler interface for Matchers.
-func (m *Matchers) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var lines []string
-	if err := unmarshal(&lines); err != nil {
-		return err
-	}
-	for _, line := range lines {
-		pm, err := labels.ParseMatchers(line)
-		if err != nil {
-			return err
-		}
-		*m = append(*m, pm...)
-	}
-	sort.Sort(labels.Matchers(*m))
-	return nil
-}
-
-// MarshalYAML implements the yaml.Marshaler interface for Matchers.
-func (m Matchers) MarshalYAML() (interface{}, error) {
-	result := make([]string, len(m))
-	for i, matcher := range m {
-		result[i] = matcher.String()
-	}
-	return result, nil
-}
-
-// UnmarshalJSON implements the json.Unmarshaler interface for Matchers.
-func (m *Matchers) UnmarshalJSON(data []byte) error {
-	var lines []string
-	if err := json.Unmarshal(data, &lines); err != nil {
-		return err
-	}
-	for _, line := range lines {
-		pm, err := labels.ParseMatchers(line)
-		if err != nil {
-			return err
-		}
-		*m = append(*m, pm...)
-	}
-	sort.Sort(labels.Matchers(*m))
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaler interface for Matchers.
-func (m Matchers) MarshalJSON() ([]byte, error) {
-	if len(m) == 0 {
-		return nil, nil
-	}
-	result := make([]string, len(m))
-	for i, matcher := range m {
-		result[i] = matcher.String()
-	}
-	return json.Marshal(result)
 }