+++ pkg/alertmanager/config/config_test.go
@@ -152,103 +152,6 @@ receivers:
 
 }
 
-func TestMuteTimeExists(t *testing.T) {
-	in := `
-route:
-    receiver: team-Y
-    routes:
-    -  match:
-        severity: critical
-       mute_time_intervals:
-       - business_hours
-
-receivers:
-- name: 'team-Y'
-`
-	_, err := Load(in)
-
-	expected := "undefined time interval \"business_hours\" used in route"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%q", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%q\ngot:\n%q", expected, err.Error())
-	}
-
-}
-
-func TestMuteTimeHasName(t *testing.T) {
-	in := `
-mute_time_intervals:
-- name:
-  time_intervals:
-  - times:
-     - start_time: '09:00'
-       end_time: '17:00'
-
-receivers:
-- name: 'team-X-mails'
-
-route:
-  receiver: 'team-X-mails'
-  routes:
-  -  match:
-      severity: critical
-     mute_time_intervals:
-     - business_hours
-`
-	_, err := Load(in)
-
-	expected := "missing name in mute time interval"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%q", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%q\ngot:\n%q", expected, err.Error())
-	}
-
-}
-
-func TestMuteTimeNoDuplicates(t *testing.T) {
-	in := `
-mute_time_intervals:
-- name: duplicate
-  time_intervals:
-  - times:
-     - start_time: '09:00'
-       end_time: '17:00'
-- name: duplicate
-  time_intervals:
-  - times:
-     - start_time: '10:00'
-       end_time: '14:00'
-
-receivers:
-- name: 'team-X-mails'
-
-route:
-  receiver: 'team-X-mails'
-  routes:
-  -  match:
-      severity: critical
-     mute_time_intervals:
-     - business_hours
-`
-	_, err := Load(in)
-
-	expected := "mute time interval \"duplicate\" is not unique"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%q", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%q\ngot:\n%q", expected, err.Error())
-	}
-
-}
-
 func TestGroupByHasNoDuplicatedLabels(t *testing.T) {
 	in := `
 route:
@@ -319,36 +222,6 @@ receivers:
 	_, err := Load(in)
 
 	expected := "no routes provided"
-
-	if err == nil {
-		t.Fatalf("no error returned, expected:\n%q", expected)
-	}
-	if err.Error() != expected {
-		t.Errorf("\nexpected:\n%q\ngot:\n%q", expected, err.Error())
-	}
-
-}
-
-func TestRootRouteNoMuteTimes(t *testing.T) {
-	in := `
-mute_time_intervals:
-- name: my_mute_time
-  time_intervals:
-  - times:
-     - start_time: '09:00'
-       end_time: '17:00'
-
-receivers:
-- name: 'team-X-mails'
-
-route:
-  receiver: 'team-X-mails'
-  mute_time_intervals:
-  - my_mute_time
-`
-	_, err := Load(in)
-
-	expected := "root route must not have any mute time intervals"
 
 	if err == nil {
 		t.Fatalf("no error returned, expected:\n%q", expected)
@@ -446,6 +319,19 @@ receivers:
 	}
 }
 
+func TestHideConfigSecrets(t *testing.T) {
+	c, err := LoadFile("testdata/conf.good.yml")
+	if err != nil {
+		t.Fatalf("Error parsing %s: %s", "testdata/conf.good.yml", err)
+	}
+
+	// String method must not reveal authentication credentials.
+	s := c.String()
+	if strings.Count(s, "<secret>") != 13 || strings.Contains(s, "mysecret") {
+		t.Fatal("config's String method reveals authentication credentials.")
+	}
+}
+
 func TestJSONMarshal(t *testing.T) {
 	c, err := LoadFile("testdata/conf.good.yml")
 	if err != nil {
@@ -456,6 +342,23 @@ func TestJSONMarshal(t *testing.T) {
 	if err != nil {
 		t.Fatal("JSON Marshaling failed:", err)
 	}
+}
+
+func TestJSONMarshalSecret(t *testing.T) {
+	test := struct {
+		S Secret
+	}{
+		S: Secret("test"),
+	}
+
+	c, err := json.Marshal(test)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// u003c -> "<"
+	// u003e -> ">"
+	require.Equal(t, "{\"S\":\"\\u003csecret\\u003e\"}", string(c), "Secret not properly elided.")
 }
 
 func TestMarshalSecretURL(t *testing.T) {
@@ -646,9 +549,7 @@ func TestEmptyFieldsAndRegex(t *testing.T) {
 	var expectedConf = Config{
 
 		Global: &GlobalConfig{
-			HTTPConfig: &commoncfg.HTTPClientConfig{
-				FollowRedirects: true,
-			},
+			HTTPConfig:      &commoncfg.HTTPClientConfig{},
 			ResolveTimeout:  model.Duration(5 * time.Minute),
 			SMTPSmarthost:   HostPort{Host: "localhost", Port: "25"},
 			SMTPFrom:        "alertmanager@example.org",
@@ -725,21 +626,6 @@ func TestEmptyFieldsAndRegex(t *testing.T) {
 
 	if !reflect.DeepEqual(configGot, configExp) {
 		t.Fatalf("%s: unexpected config result: \n\n%s\n expected\n\n%s", "testdata/conf.empty-fields.yml", configGot, configExp)
-	}
-}
-
-func TestGlobalAndLocalHTTPConfig(t *testing.T) {
-	config, err := LoadFile("testdata/conf.http-config.good.yml")
-	if err != nil {
-		t.Fatalf("Error parsing %s: %s", "testdata/conf-http-config.good.yml", err)
-	}
-
-	if config.Global.HTTPConfig.FollowRedirects {
-		t.Fatalf("global HTTP config should not follow redirects")
-	}
-
-	if !config.Receivers[0].SlackConfigs[0].HTTPConfig.FollowRedirects {
-		t.Fatalf("global HTTP config should follow redirects")
 	}
 }